+++ src/main/java/com/travel/role/domain/user/controller/UserController.java
@@ -1,22 +1,32 @@
 package com.travel.role.domain.user.controller;
 
+import javax.mail.SendFailedException;
 import javax.validation.Valid;
 
-import com.travel.role.domain.user.dto.UserPasswordModifyReqDTO;
-import com.travel.role.domain.user.dto.UserProfileDetailResDTO;
-
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.travel.role.domain.user.dto.CheckIdRequest;
+import com.travel.role.domain.user.dto.CheckIdResponse;
+import com.travel.role.domain.user.dto.ConfirmUserRequestDTO;
+import com.travel.role.domain.user.dto.ConfirmUserResponseDTO;
+import com.travel.role.domain.user.dto.NewPasswordRequestDTO;
+import com.travel.role.domain.user.dto.UserPasswordModifyReqDTO;
+import com.travel.role.domain.user.dto.UserProfileDetailResDTO;
 import com.travel.role.domain.user.dto.UserProfileModifyReqDTO;
 import com.travel.role.domain.user.dto.UserProfileResponseDTO;
+import com.travel.role.domain.user.dto.auth.SignUpRequestDTO;
+import com.travel.role.domain.user.dto.auth.SignUpResponseDTO;
 import com.travel.role.domain.user.service.UserService;
+import com.travel.role.global.auth.service.AuthService;
 import com.travel.role.global.auth.token.UserPrincipal;
 
 import lombok.RequiredArgsConstructor;
@@ -26,6 +36,7 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api")
 public class UserController {
 	private final UserService userService;
+	private final AuthService authService;
 
 	@GetMapping("/basic-profile")
 	public ResponseEntity<UserProfileResponseDTO> basicProfile(@AuthenticationPrincipal UserPrincipal userPrincipal) {
@@ -79,5 +90,28 @@ public class UserController {
 	public void deleteProfileImage(@AuthenticationPrincipal UserPrincipal userPrincipal){
 
 		userService.deleteProfileImageUrl(userPrincipal.getEmail());
+	}
+
+	@PostMapping("/signup")
+	public SignUpResponseDTO signUp(@RequestBody @Valid SignUpRequestDTO signUpRequestDTO) {
+		return authService.signUp(signUpRequestDTO);
+	}
+
+	@PostMapping("/find-id")
+	public ResponseEntity<ConfirmUserResponseDTO> confirmId(@RequestBody @Valid ConfirmUserRequestDTO confirmUserRequestDTO) {
+		ConfirmUserResponseDTO result = authService.findId(confirmUserRequestDTO);
+		return ResponseEntity.ok().body(result);
+	}
+
+	@PostMapping("/confirm-id")
+	public ResponseEntity<CheckIdResponse> confirmId(@RequestBody @Valid CheckIdRequest checkIdRequest) {
+		CheckIdResponse result = authService.confirmId(checkIdRequest);
+		return ResponseEntity.ok(result);
+	}
+
+	@PostMapping("/new-password")
+	public ResponseEntity<?> newPassword(@RequestBody NewPasswordRequestDTO dto) throws SendFailedException {
+		authService.changePassword(dto);
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 }
